+++ CHANGELOG.md
@@ -11,6 +11,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 ## Unreleased ([details][unreleased changes details])
 - #3151 - New ContentSync utility
 
+#3170 - Added a new MCP tool to bulk tag AEM content pages via an Excel file input.
+
+
+## Fixed
+
+- #3040 - Fixed bug where namespaced multi-fields would have the namespace 2 times
+- #3140 - Fixed issue where malformed MCP process nodes can cause a NPE that breaks the entire MPC reporting UI. Now displays more friendly values in UI to help remove the invalid nodes.
+- #3150 - Support for case-insensitive redirect rules ( [NC] flag equivalent of apache)
+- #3138 - Re-arrange action removes data from redirect node
+  
 ## 6.0.14 - 2023-07-11
 
 ## Fixed