+++ CHANGELOG.md
@@ -9,8 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 <!-- Keep this up to date! After a release, change the tag name to the latest release -->- 
 
 ## Unreleased ([details][unreleased changes details])
-- #3147 - Fixed setting initial content-type when importing CFs from a spreadsheet
 
+## 6.1.0 - 2023-09-08
 
 ## Added
 
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 ## Fixed
 
+- #3147 - Fixed setting initial content-type when importing CFs from a spreadsheet
 - #3040 - Fixed bug where namespaced multi-fields would have the namespace 2 times
 - #3140 - Fixed issue where malformed MCP process nodes can cause a NPE that breaks the entire MPC reporting UI. Now displays more friendly values in UI to help remove the invalid nodes.
 - #3150 - Support for case-insensitive redirect rules ( [NC] flag equivalent of apache)